requirements.txt
playwright==1.37.0
pytest==7.4.0
pytest-playwright==0.4.2
python-dotenv==1.0.0
allure-pytest==2.13.2